lsp/c: set compilationDatabaseDirectory
ccls relies on a JSON compilation database specified as compile_commands.json.
npass need this after the switch to meson build to locate header files.
for meson projects this assumes the name of the build directory is "/build".
as of writing this, the project root directory is detected using the presence of ".git".

https://git.sinanmohd.com/npass/commit/?id=4ff6158a6cd29e87222387870e6f01aca41b8183